feat: optimize file name, path
