+++ Android/app/src/main/java/com/pjh/linkument_android_app/feature/firststep/LoginUiState.kt
@@ -1,0 +1,14 @@
+package com.pjh.linkument_android_app.feature.firststep
+
+sealed interface LoginUiState {
+    object None : LoginUiState
+
+    object Loading : LoginUiState
+
+    object LoggedIn : LoginUiState
+
+    data class Error(
+        // FIXME temporary. to non-null
+        val throwable: Throwable?
+    ) : LoginUiState
+}
