+++ Android/app/src/main/java/com/pjh/linkument_android_app/feature/firststep/LoginViewModel.kt
@@ -26,16 +26,3 @@ class LoginViewModel @Inject constructor(
         }
     }
 }
-
-sealed interface LoginUiState {
-    object None : LoginUiState
-
-    object Loading : LoginUiState
-
-    object LoggedIn : LoginUiState
-
-    data class Error(
-        // FIXME temporary. to non-null
-        val throwable: Throwable?
-    ) : LoginUiState
-}
