+++ Android/app/src/main/java/com/pjh/linkument_android_app/feature/firststep/SignUpViewModel.kt
@@ -36,19 +36,4 @@ class SignUpViewModel @Inject constructor(
             }
         }
     }
-
-    sealed interface SignUpUiState {
-        object None : SignUpUiState
-
-        object Loading : SignUpUiState
-
-        object CheckValid : SignUpUiState
-
-        object Success : SignUpUiState
-
-        data class Error(
-            // FIXME temporary. to non-null
-            val throwable: Throwable?
-        ) : SignUpUiState
-    }
 }
